requirements.txt
torchvision~=0.12.0
Pillow~=9.0.1
opencv-python~=4.5.5.64
numpy~=1.22.3
torch~=1.11.0